+++ gaupol/enums.py
@@ -19,7 +19,7 @@
 
 import aeidon
 
-from aeidon.i18n import _
+from aeidon.i18n import __
 from gi.repository import Gtk
 
 __all__ = [
@@ -34,40 +34,40 @@ __all__ = [
 class FieldNumber(aeidon.EnumerationItem):
     is_position = False
     is_text = False
-    label = _("No.")
-    tooltip = _("Subtitle number")
+    label = __("No.")
+    tooltip = __("Subtitle number")
 
 class FieldStart(aeidon.EnumerationItem):
     is_position = True
     is_text = False
-    label = _("Start")
-    tooltip = _("Start position")
+    label = __("Start")
+    tooltip = __("Start position")
 
 class FieldEnd(aeidon.EnumerationItem):
     is_position = True
     is_text = False
-    label = _("End")
-    tooltip = _("End position")
+    label = __("End")
+    tooltip = __("End position")
 
 class FieldDuration(aeidon.EnumerationItem):
     is_position = True
     is_text = False
     # TRANSLATORS: 'Dur.' is short for duration. It is used in the header
     # of a column that contains numbers five digits wide.
-    label = _("Dur.")
-    tooltip = _("Duration")
+    label = __("Dur.")
+    tooltip = __("Duration")
 
 class FieldMainText(aeidon.EnumerationItem):
     is_position = False
     is_text = True
-    label = _("Text")
-    tooltip = _("Text")
+    label = __("Text")
+    tooltip = __("Text")
 
 class FieldTranslationText(aeidon.EnumerationItem):
     is_position = False
     is_text = True
-    label = _("Translation")
-    tooltip = _("Translation")
+    label = __("Translation")
+    tooltip = __("Translation")
 
 fields = aeidon.Enumeration()
 fields.NUMBER = FieldNumber()
@@ -79,10 +79,10 @@ fields.TRAN_TEXT = FieldTranslationText()
 
 
 class LengthUnitChar(aeidon.EnumerationItem):
-    label = _("characters")
+    label = __("characters")
 
 class LengthUnitEm(aeidon.EnumerationItem):
-    label = _("ems")
+    label = __("ems")
 
 length_units = aeidon.Enumeration()
 length_units.CHAR = LengthUnitChar()